requirements.txt
crewai
langchain
langchain_openai
streamlit
pandas
plotly
python-dotenv
